requirements.txt
flask 
waitress 
pytz
python-dotenv 
fastbook 
